feat: added get stateby id

+++ src/services/country-service.ts
@@ -1,3 +1,4 @@
+import mongoose from 'mongoose';
 import { ICountry } from '../interfaces/country.interface';
 import { Country } from '../models/country';
 import { CountryInput } from '../types/country';
@@ -21,6 +22,9 @@ export const getCountryByIdService = async (
 };
 
 export const deleteCountryService = async (id: string): Promise<void> => {
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    throw new Error('Invalid ID format');
+  }
   await Country.findByIdAndDelete(id);
   return;
 };
@@ -29,6 +33,9 @@ export const updateCountryService = async (
   id: string,
   newName: string,
 ): Promise<ICountry | null> => {
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    throw new Error('Invalid ID format');
+  }
   const updatedCountry = await Country.findByIdAndUpdate(
     id,
     { name: newName },
